Have yield return engine_event instead of raw coroutine
This allows for an internal unsafe_yield() which will
call coroutine.resume() directly from internal engine
supported yield functions.

This allows for an external yield() which now co_awaits
the event, and then event upon being set will correctly
resume the awaiting coroutine on the engine thread for
the user.