Fix wrong library naming in pkg-config file (#190)
The cmake configuration for generating the .pc file for pkg-config was
mistakenly prepending a 'lib' prefix. As a result, consumers of the
pkg-config file would try to link to 'liblibcoro' instead of the correct
'libcoro'. This commit adjusts the CMakeLists.txt to strip the 'lib'
prefix from the ${PROJECT_NAME} before passing it to the template for
.pc file generation. This ensures that the resulting .pc file references
the library correctly as '-lcoro', aligning with the actual name of the
library.