Use std::shared_ptr for coro::net::ssl_context
Previously this was a raw pointer to allow for the user to pass in a
global ssl context across various tcp_clients... this is pretty unsafe
and I previously moved away from this model for the coro::io_scheduler
handle that gets passed into the coro::net::tcp_client.  To be
consistent and safe this pointer is being moved to an std::shared_ptr.

Closes #204