Calling coro::sync_wait with coro::ring_buffer::consume returns default constructed objects for complex return values (#244)
Running the test that replicates the bug via valgrind or asan it shows
that the compiler is calling sync_wait()'s promise's destructor before
moving the complex object return value. This causes the object to be
destructed and then the final move out is on deleted (use after free)
memory causing the object to be in a bad empty state.

To resolve this for now a double move has been introduced to move the
complex object off the promise object and onto the sync_wait() function
callstack. This seems to keep the object alive and not be destructed
when sync_wait() finally returns.

Other solutions could be to heap allocate the promise() object or even
the return_type but this is probably more expensive than a double move,
but will vary on use case.

Closes #242