coro::when_any (#298)
Adds a new construct that will return the first task's result upon its
completion. All other task results will be discarded/detached/orphaned.

There is two ways to currently invoke when_any, one with a
std::stop_token that will signal to the other tasks that a task has
already completed, this requires the user to check for that stop token
requesting a stop, it is not automatic.

The other way is fire and forget, all tasks will be required to complete
but only the first tasks result will be used. This method isn't
particularly recommended but the API is available in the case where a
stop token isn't required.

EMSCRIPTEN does not support std::stop_source|token so this new feature
is currently disabled on that platform, I do not want to shim it in.

Closes #279