test ring_buffer hang (#327)
Changed ring buffer internals to be atomics, this really shouldn't make any difference since its behind a std::unique_lock/std::mutex, however I the `size()` function would return an unprotected/non-atomic read from the m_size field, which was reported as a data race in one of the tests.

* Renamed to shutdown() <- I would like to start being consistent on any structure that has infinite waits and requires a shutdown to follow this pattern.
* Added shutown_drain(executor_type& e) For a task based version that will yield/spin until all work is completed and then shutdown.

This may or may not fix the hang, but proceeding since the tests seem to be stable after several runs.