ring_buffer: add max_size and full functions
Add a max_size() function which returns the maximum size of the
ring_buffer. This returns the num_elements template parameter.

Add a full() function which checks if the size is equal to max_size.
This is useful in cases where code may not want to block on a produce()
call.