CMake project version now pulled from git describe
This change will now dynamically set the project VERSION in cmake to the
latest git describe tag released.

Now requires find_package(Git) to properly get the latest tag.

The tag is in the format of "v<n>.<n>.<n>-<hash>[-dirty]". This must be
parsed into a semver format of "<n>.<n>.<n>" for cmake's VERSION to
accept it.

Closes #381