coro::semaphore uses coro::mutex internally
There has been a semaphore test using a thread_pool that constantly
hangs and causes test failures. After reviewing the code this is likely
due to how the m_counter on the semaphore is handled. To remove the
ambiguity of incrementing and decrementing in a coroutine this class now
uses a coro::mutex to correctly acquire and release without going over
or under 0/max_value bounds.

Consequences of this change:
* coro::semaphore::release() is now a task and must be awaited

Closes #388