Avoid reference cycles in thread_pool and io_scheduler
Previously, the background threads in a thread_pool and/or io_scheduler
ran with a lambda capturing a shared pointer to the relevant instance. This
caused a reference cycle such that the refcount of the shared pointer
handed back to the user would never go to zero unless they explicitly
called shutdown().

To fix this, just capture a raw pointer, this is safe because the dtor for
the pool object joins the background threads and hence they can never see a
dangling pointer.

Fixes #396.