TCP client/server hold reference to shared_ptr to io_scheduler
Previously, due a reference cycle, the io_scheduler dtor would never
be called unless the user explicitly shutdown the io_scheduler
instance. Having broken that reference cycle the dtor does now run
when the scheduler goes out of scope. However, if we created TCP
client/server objects using an io_scheduler they would hold a strong
reference to the scheduler. If we then spawn a background task for
polling it can happen that the last strong reference to the
io_scheduler is held by the background polling thread and when it is
then destructed we end up joining this background thread with itself.

To fix this, only hold references to the io_scheduler in the
client/server objects.