coro::semaphore uses coro::mutex during shutdown
If there are multiple in-flight tasks acquiring a semaphore while another
task is calling shutdown, it was possible for a race to occur between setting the
shutdown flag and popping all awaiters to resume them and the in-flight
acquire suspending itself and adding itself to the awaiter list.
Consequently the in-progress suspending awaiter would get lost and never
wake.

Since both acquire and release already grab the internal mutex we can solve
the problem by doing the same in shutdown such that flipping the shutdown
switch and popping the awaiters is atomic. It now cannot be the case that
an acquire mid-suspend can add itself to awaiter list after the shutdown
flag is flipped.

As a consequence, shutdown is now also a coroutine that must be awaited.