Executor types remove circular refs
The executors are circular ref'ing with the std::shared_ptr being copied
into worker threads for coro::thread_pool and coro::io_scheduler (spawn
mode).

I'm leaving all other construct types that take an executor by
shared_ptr to continue to do so, its important that the executors they
run on do not destruct until all the tasks are completed, so the user
needs to take care to destroy objects in a meaningful way so the
executor is the last object to destruct.