Switch to std::unique_ptr
I think shared_ptr usage is just flawed since it means that any
lingering coroutine could destruct the executors. This will force the
user to make sure it outlives any coroutines at the cost of possibly
runtime nullptr references if they are going to use it past the lifetime
of the executor.