Fix deadlock in ring_buffer and queue shutdown_drain
If a ring_buffer (queue) produces an element such that the buffer is
non-empty and then calls shutdown_drain we enter a draining loop where we
yield until the ring buffer is empty. If, meanwhile, the consumer side of
the buffer decides to forcibly shutdown without consuming the produced
value (for example because it has received an exception that should be
propagated and therefore wants to wake and shutdown the producer) then we
never exit the yield-until-empty loop.

To fix this only loop while the buffer is both non-empty and we are still
in the draining state.

The produced value(s) that are stored in the buffer are now never
processed which I think is the right semantics since the consumer has
explicitly said "I can't do anything with this".

- Fixes #401