Remove potential for false-positive fail in task_container test
Since we detach our (fast to run) task, it might complete on the
thread_pool thread before we get time-sliced in to check its size on the
main thread, leading to a false positive failure in that the container is
already empty.

To fix this, use an event to ensure the task does not complete until we've
checked the container's size.