Store raw pointer in udp::peer and implement copy/move-assignment (#403)
* Store raw pointer in udp::peer and implement copy/move-assignment

Previously the peer struct held a reference member that could not be
reseated and so the defaulted copy/move-assignments were implicitly
deleted. We want these objects to be moveable and copyable so instead store
a raw pointer.

* Remove potential for false-positive fail in task_container test

Since we detach our (fast to run) task, it might complete on the
thread_pool thread before we get time-sliced in to check its size on the
main thread, leading to a false positive failure in that the container is
already empty.

To fix this, use an event to ensure the task does not complete until we've
checked the container's size.