condition_variable: mutual exclusion for notify/wait
To avoid lost wakeups, use an internal mutex to protect modification of the
awaiter list.

Now, while we are notifying, if another notification task arrives it must
wait for our modification of the awaiter list to complete. Hence waiters
that were not ready and are pushed back on the list have a chance to be
woken up by the next notification.

- Closes #398